Organizing test case assertions.
