+++ src/Illuminate/Foundation/Testing/TestCase.php
@@ -8,7 +8,6 @@ use PHPUnit_Framework_TestCase;
 abstract class TestCase extends PHPUnit_Framework_TestCase
 {
     use InteractsWithContainer,
-        AssertionsTrait,
         MakesHttpRequests,
         ImpersonatesUsers,
         InteractsWithConsole,
